BACKPORT: kbuild: allow to use GCC toolchain not in Clang search path
When using a GCC cross toolchain which is not in a compiled in
Clang search path, Clang reverts to the system assembler and
linker. This leads to assembler or linker errors, depending on
which tool is first used for a given architecture.

It seems that Clang is not searching $PATH for a matching
assembler or linker.

Make sure that Clang picks up the correct assembler or linker by
passing the cross compilers bin directory as search path.

This allows to use Clang provided by distributions with GCC
toolchains not in /usr/bin.

Link: https://github.com/ClangBuiltLinux/linux/issues/78
Signed-off-by: Stefan Agner <stefan@agner.ch>
Reviewed-and-tested-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Link: https://git.kernel.org/linus/ef8c4ed9db80261f397f0c0bf723684601ae3b52
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Change-Id: Ide6491185fa73f0458fa949294201fe5b98a0186